pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "oo-bin"
description = "Outside Open bin scripts"
authors = [
	{ name = "Andy Rusch", email = "arusch@outsideopen.com" },
	{ name = "David Lundgren", email = "dlundgren@outsideopen.com" },
	{ name = "Greg Lawler", email = "greg@outsideopen.com" },
	{ name = "Tjaart van der Walt", email = "tjaart@outsideopen.com" },
]
readme = "README.md"
requires-python = ">=3.7"
dynamic = ["version"]

dependencies = [
	"click",
	"colorama",
	"dnspython",
	"icmplib",
	"mozfile",
	"mozprofile",
	"namegenerator",
	"prettytable",
	"progress",
	"pyOpenSSL",
	"pyperclip",
	"python-whois",
	"pyxdg",
	"requests",
	"sentry-sdk",
	"singleton-decorator",
	"tabulate",
	"tomli >= 1.1.0 ; python_version < '3.11'",
	"wslPath",
]

[project.urls]
Home = "https://github.com/outsideopen/oo-bin"

[project.optional-dependencies]
test = ["pytest >=7.3.1", "pytest-mock >=3.10.0"]

[project.scripts]
oo = "oo_bin.main:main"

[tool.isort]
profile = "black"
